Persist mutations to res.locals
Adds persistence to res.locals by using Object constructor syntax rather
than Object literal syntax. This returns an object reference and means
that when passed to a function the reference is updated. As this
reference is avilable outside the middleware being tested it means that
changes to res.locals can be tested.